Split PDC server routes into multiple files
